Upgraded solution, updated fwk, refactored Resolve, moved ProxyDummyEcho out of the Logic file and implemented a (shy) CONNECT.
